Adição de registro e authenticação com jwt para proteger as rotas de acesso
Adiçao de gerador de usuario(apenas para administrador)

 links úteis:

- [Tipos de Autenticação](https://www.alura.com.br/artigos/tipos-de-autenticacao)
- [Introdução ao JSON Web Tokens (JWT)](https://jwt.io/introduction)
- [O que é JSON Web Tokens (JWT)?](https://www.alura.com.br/artigos/o-que-e-json-web-tokens)

+++ src/main/java/com/stock/flex/resource/StockResource.java
@@ -1,5 +1,6 @@
 package com.stock.flex.resource;
 
+import com.stock.flex.entity.Person;
 import com.stock.flex.entity.StockEntity;
 import com.stock.flex.repository.StockRepository;
 import com.stock.flex.resource.request.StockRequest;
@@ -8,50 +9,100 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
+
 
 @RestController
 @RequestMapping("/stock")
 public class StockResource {
 
     @Autowired
-    StockRepository repository;
+StockRepository repository;
 
     @PostMapping
     @Transactional
-    public ResponseEntity<StockResponse> create(@RequestBody StockRequest request) {
-        StockEntity savedStock = repository.save(new StockEntity(request));
+    public ResponseEntity<StockResponse> create(
+            @RequestBody StockRequest request,
+            @AuthenticationPrincipal Person userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        StockEntity newStock = new StockEntity(request);
+        newStock.setUser(userSpringSecurity);
+
+        StockEntity savedStock = repository.save(newStock);
+
         StockResponse response = new StockResponse(savedStock);
+
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     @GetMapping
-    public ResponseEntity<List<StockResponse>> get() {
-        var stock = repository.findAll().stream().map(StockResponse::new).toList();
-        return ResponseEntity.ok(stock);
-    }
+    @PreAuthorize("hasRole('ROLE_USER')")
+    public ResponseEntity<List<StockResponse>> get(
+            @AuthenticationPrincipal Person userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
 
-    @GetMapping("/{id}")
-    public ResponseEntity getById(@PathVariable UUID id) {
-        var stock = repository.getById(id);
-        return ResponseEntity.ok(new StockResponse(stock));
+        List<StockEntity> stocks = repository.findByUser(userSpringSecurity);
+
+        List<StockResponse> stockResponses = stocks.stream().map(StockResponse::new).toList();
+
+        return ResponseEntity.ok(stockResponses);
     }
 
     @PutMapping("/{id}")
     @Transactional
-    public ResponseEntity<StockResponse> update(@PathVariable UUID id, @RequestBody StockRequest request) {
-        var stock = repository.getReferenceById(id);
+    public ResponseEntity<StockResponse> update(
+            @PathVariable UUID id,
+            @RequestBody StockRequest request,
+            @AuthenticationPrincipal Person userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        StockEntity stock = repository.getById(id);
+
+        if (!stock.getUser().equals(userSpringSecurity)) {
+            throw new Exception("Acesso negado! Você não tem permissão para atualizar este estoque.");
+        }
+
         stock.updateInfo(request);
+
         return ResponseEntity.ok(new StockResponse(stock));
     }
 
     @DeleteMapping("/{id}")
     @Transactional
-    public ResponseEntity delete(@PathVariable UUID id) {
+    public ResponseEntity delete(
+            @PathVariable UUID id,
+            @AuthenticationPrincipal Person userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        StockEntity stock = repository.getById(id);
+
+        if (!stock.getUser().equals(userSpringSecurity)) {
+            throw new Exception("Acesso negado! Você não tem permissão para excluir este estoque.");
+        }
+
         repository.deleteById(id);
+
         return ResponseEntity.ok().build();
     }
+
+
 }
